+++ src/index.js
@@ -1,33 +1,20 @@
 import ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import { createStore, applyMiddleware } from "redux";
 import React from "react";
+import { Provider } from "react-redux";
+// v2.8.1 React Router.
+import { Router, Route, IndexRoute, hashHistory } from "react-router";
+import store from "./store";
 import App from "./components/App";
-import { promiseMiddleware } from "./middleware";
-
-const defaultState = {
-  appName: "Meowdium",
-  articles: null
-};
-
-const reducer = function(state = defaultState, action) {
-  switch (action.type) {
-    case "HOME_PAGE_LOADED":
-      return {
-        ...state,
-        articles: action.payload.articles,
-        articleCount: action.payload.articleCount
-      };
-    default:
-      return state;
-  }
-};
-
-const store = createStore(reducer, applyMiddleware(promiseMiddleware));
+import Home from "./components/Home";
 
 ReactDOM.render(
   <Provider store={store}>
-    <App />
+    <Router history={hashHistory}>
+      <Route path="/" component={App}>
+        <IndexRoute component={Home} />
+        {/* <Route path="/login" component={Login}/> */}
+      </Route>
+    </Router>
   </Provider>,
   document.getElementById("root")
 );
